Fixed typos and clarified empty reloids
Removed "Note that" regarding empty relative OIDs. It is not a note; it
is a requirement of X.680. And we are explicitly permitting empty
relative OIDs. I always thought that empty relative OIDs are potentially
useful, so I am glad that they are allowed in CBOR. (Note that there is
overlap in semantics between empty relative OID, null, and undefined
special values.)
Fixed two typos.